+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.0
 - (**BREAK**) changed `curAnim`, `width` and `height` by `sprite()` are changed to methods `curAnim()`, `width()` and `height()`
 - (**BREAK**) removed `velY`, `curPlatform` and `maxVel` fields by `body()`
+- fixed `dt()` surge on page visibility change (#20)
 
 # 0.2.0
 - (**BREAK**) removed `aseSpriteSheet` conf field from `loadSprite(name, src, conf)`
@@ -9,9 +10,9 @@
 - added area component generation for `text()`
 - added `noArea` to conf field of `sprite()`, `rect()` and `text()`, allowing to disable auto area component generation
 - added a `quad` field to sprite comp creation config `sprite(id, { quad: quad(0, 0, 0.5, 0.5) })`
-- fixed `resolve()` not working if the obj also has `solid`, so it does not check for itself
+- fixed `resolve()` not working if the obj also has `solid`, so it does not check for itself (#8)
 - `mousePos()` accepts a layer argument, which returns the mouse position affected by camera transform if that layer is not `camIgnore()`-ed
-- fixed camera position getting calculated before completing every object's update
+- fixed camera position getting calculated before completing every object's update (#14)
 - added `camShake()` for built in camera shake
 - fixed some cases `on("grounded", f)` called multiple times when moving on a smooth platform
 - added `revery()` to iterate objects in reverse order